1st step to limitation

+++ src/Solver/Solver.cpp
@@ -12,7 +12,7 @@ Solver::Solver( Mesh2D& msh, Problem &prb, Flux& flx) : mesh(msh),problem(prb),f
 
 void Solver::write(ostream& writer, const vector<numvector<double,5*nShapes>>& coeffs) const
 {
-    for (int k = 0; k < coeffs.size(); ++k)
+    for (size_t k = 0; k < coeffs.size(); ++k)
     {
         for (int i = 0; i < 5*nShapes; ++i)
             writer << coeffs[k][i] << ' ';
